[MOVE/#199] core 모듈 common 내부 모듈 추가
