--- core/build.gradle.kts
@@ -1,7 +1,0 @@
-plugins {
-    id("kr.genti.androidCompose")
-}
-
-android {
-    namespace = "kr.genti.core"
-}
