+++ core/common/build.gradle.kts
@@ -1,0 +1,7 @@
+plugins {
+    id("kr.genti.androidCompose")
+}
+
+android {
+    namespace = "kr.genti.core.common"
+}
