+++ core/common/src/main/java/kr/genti/common/base/BaseFragment.kt
