+++ core/common/src/main/java/kr/genti/common/extension/IntExt.kt
@@ -1,0 +1,9 @@
+package kr.genti.common.extension
+
+import android.content.Context
+import android.util.TypedValue
+
+fun Int.dpToPx(context: Context): Int {
+    val metrics = context.resources.displayMetrics
+    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this.toFloat(), metrics).toInt()
+}
