+++ core/common/src/main/java/kr/genti/common/extension/IntentExt.kt
@@ -1,0 +1,14 @@
+package kr.genti.common.extension
+
+import android.content.Intent
+import android.os.Build
+
+fun <T> Intent.getParcelable(
+    name: String,
+    clazz: Class<T>,
+): T? =
+    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
+        getParcelableExtra(name, clazz)
+    } else {
+        getParcelableExtra(name)
+    }
