+++ core/common/src/main/java/kr/genti/common/extension/StringExt.kt
@@ -1,0 +1,22 @@
+package kr.genti.common.extension
+
+import java.text.BreakIterator
+
+fun String?.isJsonObject(): Boolean = this?.startsWith("{") == true && this.endsWith("}")
+
+fun String?.isJsonArray(): Boolean = this?.startsWith("[") == true && this.endsWith("]")
+
+fun String.getGraphemeLength(): Int {
+    val breakIterator: BreakIterator = BreakIterator.getCharacterInstance()
+
+    breakIterator.setText(this)
+
+    var count = 0
+    while (breakIterator.next() != BreakIterator.DONE) {
+        count++
+    }
+
+    return count
+}
+
+fun String.breakLines(): String = this.replace(Regex("\\s+"), " ")
