+++ core/common/src/main/java/kr/genti/common/extension/UriExt.kt
@@ -1,0 +1,26 @@
+package kr.genti.common.extension
+
+import android.content.ContentResolver
+import android.net.Uri
+import android.provider.OpenableColumns
+
+fun Uri.getFileName(contentResolver: ContentResolver): String? {
+    var result: String? = null
+    if (this.scheme == "content") {
+        val cursor = contentResolver.query(this, null, null, null, null)
+        try {
+            if (cursor != null && cursor.moveToFirst()) {
+                val columnIndex = cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME)
+                if (columnIndex != -1) result = cursor.getString(columnIndex)
+            }
+        } finally {
+            cursor?.close()
+        }
+    }
+    if (result == null) {
+        result = this.path
+        val cut = result?.lastIndexOf('/')
+        if (cut != -1) result = result?.substring(cut!! + 1)
+    }
+    return result
+}
