+++ core/common/src/main/java/kr/genti/common/state/EnumUiState.kt
@@ -1,0 +1,8 @@
+package kr.genti.common.state
+
+enum class EnumUiState {
+    LOADING,
+    SUCCESS,
+    FAILURE,
+    EMPTY,
+}
