+++ core/common/src/main/java/kr/genti/common/state/UiState.kt
