+++ core/common/src/main/java/kr/genti/common/util/ItemDiffCallback.kt
@@ -1,0 +1,18 @@
+package kr.genti.common.util
+
+import androidx.recyclerview.widget.DiffUtil
+
+class ItemDiffCallback<T : Any>(
+    val onItemsTheSame: (T, T) -> Boolean,
+    val onContentsTheSame: (T, T) -> Boolean,
+) : DiffUtil.ItemCallback<T>() {
+    override fun areItemsTheSame(
+        oldItem: T,
+        newItem: T,
+    ): Boolean = onItemsTheSame(oldItem, newItem)
+
+    override fun areContentsTheSame(
+        oldItem: T,
+        newItem: T,
+    ): Boolean = onContentsTheSame(oldItem, newItem)
+}
