+++ core/common/src/main/java/kr/genti/common/util/RestartUtil.kt
@@ -1,0 +1,28 @@
+package kr.genti.common.util
+
+import android.app.ActivityOptions
+import android.content.Context
+import android.content.Intent
+
+object RestartUtil {
+    fun restartApp(
+        context: Context,
+        toastMsg: String?,
+    ) {
+        val packageManager = context.packageManager
+        val packageName = context.packageName
+        val componentName = packageManager.getLaunchIntentForPackage(packageName)?.component
+        Intent.makeRestartActivityTask(componentName).apply {
+            if (toastMsg != null) putExtra("TOAST_MSG", toastMsg)
+            context.startActivity(
+                this,
+                ActivityOptions.makeCustomAnimation(
+                    context,
+                    0,
+                    0,
+                ).toBundle(),
+            )
+        }
+        Runtime.getRuntime().exit(0)
+    }
+}
