+++ core/common/src/main/res/anim/fade_in.xml
