+++ core/common/src/main/res/anim/fade_out.xml
