+++ core/common/src/main/res/values/styles.xml
