--- core/src/main/java/kr/genti/core/base/BaseBottomSheet.kt
@@ -1,40 +1,0 @@
-package kr.genti.core.base
-
-import android.app.Dialog
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.annotation.LayoutRes
-import androidx.databinding.DataBindingUtil
-import androidx.databinding.ViewDataBinding
-import com.google.android.material.bottomsheet.BottomSheetDialogFragment
-
-abstract class BaseBottomSheet<T : ViewDataBinding>(
-    @LayoutRes private val layoutRes: Int,
-) : BottomSheetDialogFragment() {
-    private var _binding: T? = null
-    protected val binding: T
-        get() = requireNotNull(_binding) { "binding object is not initialized" }
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?,
-    ): View? {
-        _binding = DataBindingUtil.inflate(inflater, layoutRes, container, false)
-        binding.lifecycleOwner = viewLifecycleOwner
-        return binding.root
-    }
-
-    override fun onCreateDialog(savedInstanceState: Bundle?): Dialog {
-        return super.onCreateDialog(savedInstanceState).apply {
-            window?.setDimAmount(0.8f)
-        }
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-}
