--- core/src/main/java/kr/genti/core/extension/IntExt.kt
@@ -1,9 +1,0 @@
-package kr.genti.core.extension
-
-import android.content.Context
-import android.util.TypedValue
-
-fun Int.dpToPx(context: Context): Int {
-    val metrics = context.resources.displayMetrics
-    return TypedValue.applyDimension(TypedValue.COMPLEX_UNIT_DIP, this.toFloat(), metrics).toInt()
-}
