--- core/src/main/java/kr/genti/core/extension/IntentExt.kt
@@ -1,14 +1,0 @@
-package kr.genti.core.extension
-
-import android.content.Intent
-import android.os.Build
-
-fun <T> Intent.getParcelable(
-    name: String,
-    clazz: Class<T>,
-): T? =
-    if (Build.VERSION.SDK_INT >= Build.VERSION_CODES.TIRAMISU) {
-        getParcelableExtra(name, clazz)
-    } else {
-        getParcelableExtra(name)
-    }
