--- core/src/main/java/kr/genti/core/extension/StringExt.kt
@@ -1,22 +1,0 @@
-package kr.genti.core.extension
-
-import java.text.BreakIterator
-
-fun String?.isJsonObject(): Boolean = this?.startsWith("{") == true && this.endsWith("}")
-
-fun String?.isJsonArray(): Boolean = this?.startsWith("[") == true && this.endsWith("]")
-
-fun String.getGraphemeLength(): Int {
-    val breakIterator: BreakIterator = BreakIterator.getCharacterInstance()
-
-    breakIterator.setText(this)
-
-    var count = 0
-    while (breakIterator.next() != BreakIterator.DONE) {
-        count++
-    }
-
-    return count
-}
-
-fun String.breakLines(): String = this.replace(Regex("\\s+"), " ")
