--- core/src/main/java/kr/genti/core/extension/UriExt.kt
@@ -1,26 +1,0 @@
-package kr.genti.core.extension
-
-import android.content.ContentResolver
-import android.net.Uri
-import android.provider.OpenableColumns
-
-fun Uri.getFileName(contentResolver: ContentResolver): String? {
-    var result: String? = null
-    if (this.scheme == "content") {
-        val cursor = contentResolver.query(this, null, null, null, null)
-        try {
-            if (cursor != null && cursor.moveToFirst()) {
-                val columnIndex = cursor.getColumnIndex(OpenableColumns.DISPLAY_NAME)
-                if (columnIndex != -1) result = cursor.getString(columnIndex)
-            }
-        } finally {
-            cursor?.close()
-        }
-    }
-    if (result == null) {
-        result = this.path
-        val cut = result?.lastIndexOf('/')
-        if (cut != -1) result = result?.substring(cut!! + 1)
-    }
-    return result
-}
