--- core/src/main/java/kr/genti/core/state/EnumUiState.kt
@@ -1,8 +1,0 @@
-package kr.genti.core.state
-
-enum class EnumUiState {
-    LOADING,
-    SUCCESS,
-    FAILURE,
-    EMPTY,
-}
