--- core/src/main/java/kr/genti/core/util/ItemDiffCallback.kt
@@ -1,18 +1,0 @@
-package kr.genti.core.util
-
-import androidx.recyclerview.widget.DiffUtil
-
-class ItemDiffCallback<T : Any>(
-    val onItemsTheSame: (T, T) -> Boolean,
-    val onContentsTheSame: (T, T) -> Boolean,
-) : DiffUtil.ItemCallback<T>() {
-    override fun areItemsTheSame(
-        oldItem: T,
-        newItem: T,
-    ): Boolean = onItemsTheSame(oldItem, newItem)
-
-    override fun areContentsTheSame(
-        oldItem: T,
-        newItem: T,
-    ): Boolean = onContentsTheSame(oldItem, newItem)
-}
