+++ data/build.gradle.kts
@@ -1,3 +1,5 @@
+import kr.genti.convention.extension.implementation
+
 plugins {
     id("kr.genti.androidLibrary")
 }
@@ -8,4 +10,9 @@ android {
 
 dependencies {
     implementation(projects.domain)
+
+    implementation(platform(libs.okhttp.bom))
+    implementation(libs.bundles.okhttp)
+    implementation(platform(libs.retrofit.bom))
+    implementation(libs.bundles.retrofit)
 }
